ACM-39327: extract SSE RBAC modules and derive cluster-scoped kinds
Split access cache and SelfSubjectRulesReview logic out of events.ts so
watch definitions own cluster scope, SSAR cache keys include API group,
and failed rules reviews can retry instead of pinning incomplete results.
