+++ backend/test/routes/events.test.ts
@@ -15,15 +15,6 @@ import {
   createWatchEventProcessor,
   listAndWatch,
   stopWatching,
-  canAccess,
-  canGetResource,
-  resetAccessCache,
-  getAccessCache,
-  cleanupAccessCache,
-  hashAccessToken,
-  ACCESS_CACHE_TTL,
-  ACCESS_CACHE_MAX_TOKENS,
-  ACCESS_CACHE_MAX_ENTRIES_PER_TOKEN,
 } from '../../src/routes/events'
 import * as serviceAccountTokenModule from '../../src/lib/serviceAccountToken'
 import type { IArgoApplication, IResource } from '../../src/resources/resource'
@@ -1412,862 +1403,6 @@ describe('events Route', () => {
       const retryTime = secondListTime - startTime
       expect(retryTime).toBeLessThan(5000)
       expect(listCallCount).toBe(1) // Only counting second list call
-    })
-  })
-
-  describe('Access Cache Cleanup', () => {
-    beforeEach(() => {
-      resetAccessCache()
-      jest.clearAllMocks()
-      process.env.CLUSTER_API_URL = 'https://api.test-cluster.com:6443'
-    })
-
-    afterEach(() => {
-      resetAccessCache()
-      delete process.env.CLUSTER_API_URL
-      nock.cleanAll()
-    })
-
-    it('should cache RBAC access check results under hashed token keys', async () => {
-      const mockToken = 'test-token-123'
-      const resource = { kind: 'Pod', apiVersion: 'v1', metadata: { namespace: 'default', name: 'test-pod' } }
-
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      const result1 = await canAccess(resource, 'get', mockToken)
-      const result2 = await canAccess(resource, 'get', mockToken)
-
-      expect(result1).toBe(true)
-      expect(result1).toBe(result2)
-      expect(getAccessCache()[mockToken]).toBeUndefined()
-      expect(getAccessCache()[hashAccessToken(mockToken)]['get:Pod:default:test-pod']).toBeDefined()
-    })
-
-    it('should use distinct cache keys per verb', async () => {
-      const mockToken = 'test-token-verb'
-      const resource = { kind: 'Pod', apiVersion: 'v1', metadata: { namespace: 'default', name: 'test-pod' } }
-
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: false } })
-
-      expect(await canAccess(resource, 'get', mockToken)).toBe(true)
-      expect(await canAccess(resource, 'list', mockToken)).toBe(false)
-
-      const tokenCache = getAccessCache()[hashAccessToken(mockToken)]
-      expect(tokenCache['get:Pod:default:test-pod']).toBeDefined()
-      expect(tokenCache['list:Pod:default:test-pod']).toBeDefined()
-    })
-
-    it('should log access checks when LOG_ACCESS is enabled', async () => {
-      const { logger } = await import('../../src/lib/logger')
-      const debugSpy = jest.spyOn(logger, 'debug').mockImplementation(() => {})
-      process.env.LOG_ACCESS = 'true'
-
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      await canAccess(
-        { kind: 'Pod', apiVersion: 'v1', metadata: { namespace: 'default', name: 'logged-pod' } },
-        'get',
-        'log-access-token'
-      )
-
-      expect(debugSpy).toHaveBeenCalledWith(
-        expect.objectContaining({ msg: 'access', allowed: true, verb: 'get', resource: 'pods' })
-      )
-
-      debugSpy.mockRestore()
-      delete process.env.LOG_ACCESS
-    })
-
-    it('should respect TTL and refetch after expiry', async () => {
-      const cache = getAccessCache()
-      const mockToken = 'test-token-ttl'
-      const tokenKey = hashAccessToken(mockToken)
-
-      cache[tokenKey] = {
-        'get:Secret:default:credentials': {
-          time: Date.now() - ACCESS_CACHE_TTL - 1000,
-          promise: Promise.resolve(true),
-        },
-      }
-
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: false } })
-
-      const result = await canAccess(
-        { kind: 'Secret', apiVersion: 'v1', metadata: { namespace: 'default', name: 'credentials' } },
-        'get',
-        mockToken
-      )
-      expect(result).toBe(false)
-    })
-
-    it('should remove stale cache entries during cleanup', () => {
-      const cache = getAccessCache()
-      const now = Date.now()
-
-      cache['token1'] = {
-        stale: { time: now - ACCESS_CACHE_TTL - 1000, promise: Promise.resolve(true) },
-        fresh: { time: now - 30000, promise: Promise.resolve(true) },
-      }
-      cache['token2'] = { 'stale-only': { time: now - ACCESS_CACHE_TTL - 5000, promise: Promise.resolve(false) } }
-
-      cleanupAccessCache()
-
-      expect(cache['token1']['stale']).toBeUndefined()
-      expect(cache['token1']['fresh']).toBeDefined()
-      expect(cache['token2']).toBeUndefined()
-    })
-
-    it('should expire subject rules and kind access caches during cleanup', async () => {
-      let rulesCalls = 0
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .twice()
-        .reply(200, () => {
-          rulesCalls++
-          return { status: { incomplete: false, resourceRules: [] } }
-        })
-
-      const start = 1_000_000
-      const nowSpy = jest.spyOn(Date, 'now').mockReturnValue(start)
-
-      await canGetResource(
-        { kind: 'ManagedCluster', apiVersion: 'cluster.open-cluster-management.io/v1', metadata: { name: 'c1' } },
-        'cache-expiry-token'
-      )
-
-      nowSpy.mockReturnValue(start + ACCESS_CACHE_TTL + 1)
-      cleanupAccessCache()
-
-      nowSpy.mockReturnValue(start + ACCESS_CACHE_TTL + 2)
-      await canGetResource(
-        { kind: 'ManagedCluster', apiVersion: 'cluster.open-cluster-management.io/v1', metadata: { name: 'c2' } },
-        'cache-expiry-token'
-      )
-
-      nowSpy.mockRestore()
-      expect(rulesCalls).toBe(2)
-    })
-
-    it('should enforce maximum token limit with LRU eviction', () => {
-      const cache = getAccessCache()
-      const now = Date.now()
-      const tokenCount = ACCESS_CACHE_MAX_TOKENS + 100
-
-      for (let i = 0; i < tokenCount; i++) {
-        cache[`token-${i}`] = {
-          'Pod:default:test': { time: now - (i / tokenCount) * 50 * 1000, promise: Promise.resolve(true) },
-        }
-      }
-
-      cleanupAccessCache()
-
-      expect(Object.keys(cache).length).toBe(ACCESS_CACHE_MAX_TOKENS)
-      expect(cache['token-0']).toBeDefined()
-      expect(cache[`token-${tokenCount - 1}`]).toBeUndefined()
-    })
-
-    it('should enforce maximum entries per token', async () => {
-      const mockToken = 'test-token-entry-cap'
-
-      nock(process.env.CLUSTER_API_URL || '')
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .times(ACCESS_CACHE_MAX_ENTRIES_PER_TOKEN + 50)
-        .reply(200, { status: { allowed: false } })
-
-      for (let i = 0; i < ACCESS_CACHE_MAX_ENTRIES_PER_TOKEN + 50; i++) {
-        await canAccess(
-          { kind: 'Pod', apiVersion: 'v1', metadata: { namespace: 'default', name: `pod-${i}` } },
-          'get',
-          mockToken
-        )
-      }
-
-      const tokenCache = getAccessCache()[hashAccessToken(mockToken)]
-      expect(Object.keys(tokenCache).length).toBeLessThanOrEqual(ACCESS_CACHE_MAX_ENTRIES_PER_TOKEN)
-    })
-  })
-
-  /**
-   * ACM-39327: restricted users must not trigger O(N) SelfSubjectAccessReviews when the SSE
-   * filter falls through after cluster-scoped list is denied. SelfSubjectRulesReview is
-   * namespaced: cache one review per token+namespace, never treat `default` as global allow.
-   */
-  describe('SelfSubjectRulesReview short-circuit (ACM-39327)', () => {
-    const apiUrl = () => process.env.CLUSTER_API_URL || ''
-    const managedCluster = (name: string) => ({
-      kind: 'ManagedCluster',
-      apiVersion: 'cluster.open-cluster-management.io/v1',
-      metadata: { name },
-    })
-    const secret = (namespace: string, name: string) => ({
-      kind: 'Secret',
-      apiVersion: 'v1',
-      metadata: { namespace, name },
-    })
-    const managedClusterInfo = (cluster: string) => ({
-      kind: 'ManagedClusterInfo',
-      apiVersion: 'internal.open-cluster-management.io/v1beta1',
-      metadata: { name: cluster, namespace: cluster },
-    })
-
-    const emptyRules = { incomplete: false, resourceRules: [] as unknown[] }
-    const secretGetInNamespace = {
-      incomplete: false,
-      resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['secrets'] }],
-    }
-    const clusterAdminRules = {
-      incomplete: false,
-      resourceRules: [{ verbs: ['*'], apiGroups: ['*'], resources: ['*'] }],
-    }
-
-    function rulesReviewNamespace(body: unknown): string {
-      let parsed = body
-      if (typeof body === 'string') {
-        try {
-          parsed = JSON.parse(body) as unknown
-        } catch {
-          return ''
-        }
-      }
-      return (parsed as { spec?: { namespace?: string } })?.spec?.namespace || ''
-    }
-
-    function nockRulesReview(replyFn: (namespace: string) => { incomplete?: boolean; resourceRules?: unknown[] }) {
-      return nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .reply(200, (_uri: string, requestBody: unknown) => ({ status: replyFn(rulesReviewNamespace(requestBody)) }))
-    }
-
-    function parseSsarResourceAttributes(body: unknown) {
-      const parsed =
-        typeof body === 'string'
-          ? (JSON.parse(body) as { spec?: { resourceAttributes?: Record<string, string | undefined> } })
-          : body
-      return (parsed as { spec?: { resourceAttributes?: Record<string, string | undefined> } })?.spec
-        ?.resourceAttributes
-    }
-
-    function nockSsarGet(matcher: (attrs: Record<string, string | undefined>) => boolean, allowed: boolean) {
-      return nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews', (body: unknown) => {
-          const attrs = parseSsarResourceAttributes(body)
-          return attrs?.verb === 'get' && matcher(attrs ?? {})
-        })
-        .reply(200, { status: { allowed } })
-    }
-
-    const namedManagedClusterRule = (name: string) => ({
-      incomplete: false,
-      resourceRules: [
-        {
-          verbs: ['get'],
-          apiGroups: ['cluster.open-cluster-management.io'],
-          resources: ['managedclusters'],
-          resourceNames: [name],
-        },
-      ],
-    })
-
-    beforeEach(() => {
-      resetAccessCache()
-      jest.clearAllMocks()
-      process.env.CLUSTER_API_URL = 'https://api.test-cluster.com:6443'
-    })
-
-    afterEach(() => {
-      resetAccessCache()
-      delete process.env.CLUSTER_API_URL
-      nock.cleanAll()
-    })
-
-    it('should deny all gets from complete empty rules without per-object SSAR', async () => {
-      nockRulesReview(() => emptyRules)
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'none-user-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should use a single rules review for many cluster-scoped gets of the same kind', async () => {
-      let rulesCalls = 0
-      nockRulesReview(() => {
-        rulesCalls++
-        return emptyRules
-      })
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .times(1)
-        .reply(200, { status: { allowed: true } })
-
-      const results = await Promise.all(
-        Array.from({ length: 500 }, (_, i) => canGetResource(managedCluster(`cluster-${i}`), 'scale-none-token'))
-      )
-
-      expect(results.every((allowed) => allowed === false)).toBe(true)
-      expect(rulesCalls).toBe(1)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should use a single rules review for many namespaced gets in the same namespace', async () => {
-      let rulesCalls = 0
-      nockRulesReview(() => {
-        rulesCalls++
-        return emptyRules
-      })
-
-      const results = await Promise.all(
-        Array.from({ length: 200 }, (_, i) =>
-          canGetResource(
-            {
-              kind: 'ManagedClusterInfo',
-              apiVersion: 'internal.open-cluster-management.io/v1beta1',
-              metadata: { name: `info-${i}`, namespace: 'acm39327-mc-01' },
-            },
-            'same-ns-none-token'
-          )
-        )
-      )
-
-      expect(results.every((allowed) => allowed === false)).toBe(true)
-      expect(rulesCalls).toBe(1)
-    })
-
-    it('should issue one rules review per namespace for a none user, without per-object SSAR', async () => {
-      const namespaces = new Set<string>()
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          namespaces.add(rulesReviewNamespace(body))
-          return true
-        })
-        .times(50)
-        .reply(200, { status: emptyRules })
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      const results = await Promise.all(
-        Array.from({ length: 50 }, (_, i) =>
-          canGetResource(managedClusterInfo(`cluster-${i}`), 'namespaced-none-token')
-        )
-      )
-
-      expect(results.every((allowed) => allowed === false)).toBe(true)
-      expect(namespaces.size).toBe(50)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('must not treat get secrets in default as access to Credentials in other namespaces', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) === 'default'
-        })
-        .reply(200, { status: secretGetInNamespace })
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) !== 'default'
-        })
-        .times(2)
-        .reply(200, { status: emptyRules })
-
-      expect(await canGetResource(secret('default', 'default-cred'), 'user1-token')).toBe(true)
-      expect(await canGetResource(secret('kube-system', 'other-cred'), 'user1-token')).toBe(false)
-      expect(await canGetResource(secret('acm39327-mc-01', 'cluster-cred'), 'user1-token')).toBe(false)
-    })
-
-    it('should allow namespaced cluster resources when the user is admin in that cluster namespace', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) === 'acm39327-mc-01'
-        })
-        .reply(200, { status: clusterAdminRules })
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) !== 'acm39327-mc-01'
-        })
-        .times(2)
-        .reply(200, { status: emptyRules })
-
-      expect(await canGetResource(managedClusterInfo('acm39327-mc-01'), 'cluster-admin-token')).toBe(true)
-      expect(await canGetResource(managedClusterInfo('other-cluster'), 'cluster-admin-token')).toBe(false)
-      expect(await canGetResource(managedCluster('acm39327-mc-01'), 'cluster-admin-token')).toBe(false)
-    })
-
-    it('should confirm named cluster-scoped resources with SSAR before allowing access', async () => {
-      nockRulesReview(() => namedManagedClusterRule('allowed-cluster'))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'allowed-cluster',
-        true
-      )
-
-      expect(await canGetResource(managedCluster('allowed-cluster'), 'partial-user-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should deny non-matching names from cluster-scoped allow-names without trusting SSRR alone', async () => {
-      nockRulesReview(() => namedManagedClusterRule('allowed-cluster'))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'other-cluster',
-        false
-      )
-
-      expect(await canGetResource(managedCluster('other-cluster'), 'partial-user-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should allow namespaced resources when rules grant unrestricted get/list/watch in that namespace', async () => {
-      nockRulesReview(() => ({
-        incomplete: false,
-        resourceRules: [{ verbs: ['get', 'list', 'watch'], apiGroups: [''], resources: ['secrets'] }],
-      }))
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: false } })
-
-      expect(await canGetResource(secret('default', 'any-secret'), 'viewer-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should confirm cluster-scoped allow-all with SSAR so default RoleBindings are not treated as cluster access', async () => {
-      nockRulesReview(() => ({
-        incomplete: false,
-        resourceRules: [
-          {
-            verbs: ['get', 'list', 'watch'],
-            apiGroups: ['cluster.open-cluster-management.io'],
-            resources: ['managedclusters'],
-          },
-        ],
-      }))
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: false } })
-
-      expect(await canGetResource(managedCluster('any-cluster'), 'default-role-token')).toBe(false)
-    })
-
-    it('should reuse kind access across API versions of the same group', async () => {
-      let rulesCalls = 0
-      nockRulesReview(() => {
-        rulesCalls++
-        return emptyRules
-      })
-
-      expect(
-        await canGetResource(
-          { kind: 'Placement', apiVersion: 'cluster.open-cluster-management.io/v1beta1', metadata: { name: 'p1' } },
-          'version-token'
-        )
-      ).toBe(false)
-      expect(
-        await canGetResource(
-          { kind: 'Placement', apiVersion: 'cluster.open-cluster-management.io/v1alpha1', metadata: { name: 'p2' } },
-          'version-token'
-        )
-      ).toBe(false)
-      expect(rulesCalls).toBe(1)
-    })
-
-    it('should confirm cluster-scoped incomplete rules with SSAR instead of trusting SSRR alone', async () => {
-      nockRulesReview(() => ({
-        incomplete: true,
-        resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['pods'] }],
-      }))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'cluster-1',
-        true
-      )
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'incomplete-user-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should deny-all when rules are empty even if incomplete is true (OpenShift none user)', async () => {
-      nockRulesReview(() => ({ incomplete: true, resourceRules: [] }))
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'openshift-none-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should fall back to SSAR when SelfSubjectRulesReview request fails', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .reply(500, { message: 'internal error' })
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'ssrr-fail-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should reuse subject rules cache across kinds in the same namespace', async () => {
-      let rulesCalls = 0
-      nockRulesReview(() => {
-        rulesCalls++
-        return emptyRules
-      })
-
-      await canGetResource(managedCluster('cluster-1'), 'shared-rules-token')
-      await canGetResource(
-        { kind: 'Placement', apiVersion: 'cluster.open-cluster-management.io/v1beta1', metadata: { name: 'p1' } },
-        'shared-rules-token'
-      )
-
-      expect(rulesCalls).toBe(1)
-    })
-
-    it('should deny-all when evaluationError is set with empty resourceRules', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .reply(200, {
-          status: {
-            incomplete: false,
-            evaluationError: 'authorizer unavailable',
-            resourceRules: [],
-          },
-        })
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'eval-error-empty-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should deny-all when rules do not grant access to the requested kind', async () => {
-      nockRulesReview(() => ({
-        incomplete: false,
-        resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['pods'] }],
-      }))
-
-      const ssarScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews')
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'unrelated-rules-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(false)
-    })
-
-    it('should fall back to namespaced list SSAR for incomplete namespaced resources', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) === 'acm39327-mc-01'
-        })
-        .reply(200, {
-          status: {
-            incomplete: true,
-            resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['secrets'] }],
-          },
-        })
-
-      const listScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews', (body: unknown) => {
-          const parsed =
-            typeof body === 'string'
-              ? (JSON.parse(body) as { spec?: { resourceAttributes?: { verb?: string; namespace?: string } } })
-              : body
-          const attrs = (parsed as { spec?: { resourceAttributes?: { verb?: string; namespace?: string } } })?.spec
-            ?.resourceAttributes
-          return attrs?.verb === 'list' && attrs?.namespace === 'acm39327-mc-01'
-        })
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedClusterInfo('acm39327-mc-01'), 'namespaced-incomplete-token')).toBe(true)
-      expect(listScope.isDone()).toBe(true)
-    })
-
-    it('should fall back to get SSAR when namespaced incomplete rules deny list access', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) === 'acm39327-mc-01'
-        })
-        .reply(200, {
-          status: {
-            incomplete: true,
-            resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['secrets'] }],
-          },
-        })
-
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews', (body: unknown) => {
-          const parsed =
-            typeof body === 'string'
-              ? (JSON.parse(body) as { spec?: { resourceAttributes?: { verb?: string; namespace?: string } } })
-              : body
-          const attrs = (parsed as { spec?: { resourceAttributes?: { verb?: string; namespace?: string } } })?.spec
-            ?.resourceAttributes
-          return attrs?.verb === 'list' && attrs?.namespace === 'acm39327-mc-01'
-        })
-        .reply(200, { status: { allowed: false } })
-
-      const getScope = nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews', (body: unknown) => {
-          const parsed =
-            typeof body === 'string'
-              ? (JSON.parse(body) as {
-                  spec?: { resourceAttributes?: { verb?: string; namespace?: string; name?: string } }
-                })
-              : body
-          const attrs = (
-            parsed as { spec?: { resourceAttributes?: { verb?: string; namespace?: string; name?: string } } }
-          )?.spec?.resourceAttributes
-          return attrs?.verb === 'get' && attrs?.namespace === 'acm39327-mc-01' && attrs?.name === 'acm39327-mc-01'
-        })
-        .reply(200, { status: { allowed: true } })
-
-      expect(await canGetResource(managedClusterInfo('acm39327-mc-01'), 'namespaced-incomplete-get-token')).toBe(true)
-      expect(getScope.isDone()).toBe(true)
-    })
-
-    it('should deny allow-names when the resource has no name', async () => {
-      nockRulesReview(() => ({
-        incomplete: false,
-        resourceRules: [
-          {
-            verbs: ['get'],
-            apiGroups: [''],
-            resources: ['secrets'],
-            resourceNames: ['named-secret'],
-          },
-        ],
-      }))
-
-      expect(
-        await canGetResource(
-          { kind: 'Secret', apiVersion: 'v1', metadata: { namespace: 'default' } },
-          'named-only-token'
-        )
-      ).toBe(false)
-    })
-  })
-
-  describe('cluster-scoped SSRR middle-ground security (ACM-39327)', () => {
-    const apiUrl = () => process.env.CLUSTER_API_URL || ''
-    const managedCluster = (name: string) => ({
-      kind: 'ManagedCluster',
-      apiVersion: 'cluster.open-cluster-management.io/v1',
-      metadata: { name },
-    })
-
-    function rulesReviewNamespace(body: unknown): string {
-      let parsed = body
-      if (typeof body === 'string') {
-        try {
-          parsed = JSON.parse(body) as unknown
-        } catch {
-          return ''
-        }
-      }
-      return (parsed as { spec?: { namespace?: string } })?.spec?.namespace || ''
-    }
-
-    function nockRulesReview(replyFn: (namespace: string) => { incomplete?: boolean; resourceRules?: unknown[] }) {
-      return nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .reply(200, (_uri: string, requestBody: unknown) => ({ status: replyFn(rulesReviewNamespace(requestBody)) }))
-    }
-
-    function nockRulesReviewStatus(
-      replyFn: (namespace: string) => {
-        incomplete?: boolean
-        evaluationError?: string
-        resourceRules?: unknown[]
-      }
-    ) {
-      return nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews')
-        .reply(200, (_uri: string, requestBody: unknown) => ({ status: replyFn(rulesReviewNamespace(requestBody)) }))
-    }
-
-    function parseSsarResourceAttributes(body: unknown) {
-      const parsed =
-        typeof body === 'string'
-          ? (JSON.parse(body) as { spec?: { resourceAttributes?: Record<string, string | undefined> } })
-          : body
-      return (parsed as { spec?: { resourceAttributes?: Record<string, string | undefined> } })?.spec
-        ?.resourceAttributes
-    }
-
-    function nockSsarGet(matcher: (attrs: Record<string, string | undefined>) => boolean, allowed: boolean) {
-      return nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectaccessreviews', (body: unknown) => {
-          const attrs = parseSsarResourceAttributes(body)
-          return attrs?.verb === 'get' && matcher(attrs ?? {})
-        })
-        .reply(200, { status: { allowed } })
-    }
-
-    const namedManagedClusterRule = (name: string) => ({
-      incomplete: false,
-      resourceRules: [
-        {
-          verbs: ['get'],
-          apiGroups: ['cluster.open-cluster-management.io'],
-          resources: ['managedclusters'],
-          resourceNames: [name],
-        },
-      ],
-    })
-
-    beforeEach(() => {
-      resetAccessCache()
-      jest.clearAllMocks()
-      process.env.CLUSTER_API_URL = 'https://api.test-cluster.com:6443'
-    })
-
-    afterEach(() => {
-      resetAccessCache()
-      delete process.env.CLUSTER_API_URL
-      nock.cleanAll()
-    })
-
-    it('should deny allow-names from a default RoleBinding when SSAR get is false', async () => {
-      nockRulesReview(() => namedManagedClusterRule('acm39327-mc-02'))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'acm39327-mc-02',
-        false
-      )
-
-      expect(await canGetResource(managedCluster('acm39327-mc-02'), 'user1-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should allow allow-names only when SSAR get confirms a real ClusterRoleBinding grant', async () => {
-      nockRulesReview(() => namedManagedClusterRule('allowed-cluster'))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'allowed-cluster',
-        true
-      )
-
-      expect(await canGetResource(managedCluster('allowed-cluster'), 'clusterrole-user-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should confirm cluster-scoped allow-all with SSAR and deny when SSAR rejects', async () => {
-      nockRulesReview(() => ({
-        incomplete: false,
-        resourceRules: [
-          {
-            verbs: ['get', 'list', 'watch'],
-            apiGroups: ['cluster.open-cluster-management.io'],
-            resources: ['managedclusters'],
-          },
-        ],
-      }))
-      const ssarScope = nockSsarGet(
-        (attrs) => attrs.group === 'cluster.open-cluster-management.io' && attrs.resource === 'managedclusters',
-        false
-      )
-
-      expect(await canGetResource(managedCluster('any-cluster'), 'default-role-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should not trust allow-names on ManagedCluster when metadata.namespace is set without SSAR confirmation', async () => {
-      nock(apiUrl())
-        .post('/apis/authorization.k8s.io/v1/selfsubjectrulesreviews', (body: unknown) => {
-          return rulesReviewNamespace(body) === 'default'
-        })
-        .reply(200, { status: namedManagedClusterRule('acm39327-mc-02') })
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'acm39327-mc-02',
-        false
-      )
-
-      expect(
-        await canGetResource(
-          {
-            kind: 'ManagedCluster',
-            apiVersion: 'cluster.open-cluster-management.io/v1',
-            metadata: { name: 'acm39327-mc-02', namespace: 'default' },
-          },
-          'user1-token'
-        )
-      ).toBe(false)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should treat SSRR evaluationError as incomplete and confirm cluster-scoped access with SSAR', async () => {
-      nockRulesReviewStatus(() => ({
-        incomplete: false,
-        evaluationError: 'webhook authorizer does not support user rule resolution',
-        resourceRules: [{ verbs: ['get'], apiGroups: [''], resources: ['pods'] }],
-      }))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'cluster-1',
-        true
-      )
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'evaluation-error-token')).toBe(true)
-      expect(ssarScope.isDone()).toBe(true)
-    })
-
-    it('should confirm any non-deny-all cluster-scoped SSRR result with SSAR, not applyKindGetAccess alone', async () => {
-      nockRulesReview(() => ({
-        incomplete: true,
-        resourceRules: [
-          {
-            verbs: ['get'],
-            apiGroups: ['cluster.open-cluster-management.io'],
-            resources: ['managedclusters'],
-            resourceNames: ['cluster-1'],
-          },
-        ],
-      }))
-      const ssarScope = nockSsarGet(
-        (attrs) =>
-          attrs.group === 'cluster.open-cluster-management.io' &&
-          attrs.resource === 'managedclusters' &&
-          attrs.name === 'cluster-1',
-        false
-      )
-
-      expect(await canGetResource(managedCluster('cluster-1'), 'incomplete-named-token')).toBe(false)
-      expect(ssarScope.isDone()).toBe(true)
     })
   })
 })
